+++ src/helper/messages.ts
@@ -1,7 +1,7 @@
 export default {
   common: {
     page: "Page <page>/<total>",
-    no_permission: "You don't have permission to use this command",
+    no_permission: "You don't have permission to use this command"
   },
   birthday: {
     user: "<user>'s birthday: <date>",
@@ -12,13 +12,14 @@ export default {
     today_bday_s: "Deem todos os parabéns aos aniversariantes de hoje:",
     already_set: "User already has a birthday set",
     not_found: "User birthday not found",
+    found_zero: "No birthdays found for the month <month>",
     set_success: "Birthday to user <user> set successfully for date <date>",
     remove_success: "User birthday removed",
     channel_set_success: "Birthday channel set successfully to <channel>",
-    servers_channel: "Server's birthday channel: <channel>",
+    servers_channel: "Server's birthday channel: <channel>"
   },
   fandom: {
-    invalid_slug: "Invalid fandom slug",
+    invalid_slug: "Invalid fandom slug"
   },
   livechart: {
     list: "<user> subscription list",
@@ -29,7 +30,7 @@ export default {
     sub_doesnt_exists: "Subscription for <link> doesn't exist",
     channel_set_success: "Anime channel set successfully to <channel>",
     sub_success: "Subscription for <link> addedd successfully",
-    unsub_success: "Subscription for <link> removed successfully",
+    unsub_success: "Subscription for <link> removed successfully"
   },
   mal: {
     type: "Type: <type>",
@@ -39,16 +40,16 @@ export default {
     related_works: "Related works",
     alternative_names: "Alternative names",
     birthday: "Birthday",
-    not_found: "No content found for that query",
+    not_found: "No content found for that query"
   },
   misc: {
-    group: "Group <index>",
+    group: "Group <index>"
   },
   sauce: {
     not_found: "No source found",
     similarity: "Similarity",
     season: "Season",
     other_names: "Other names",
-    image_not_found: "No image found to request.",
-  },
+    image_not_found: "No image found to request."
+  }
 };