+++ src/types/types.d.ts
@@ -165,38 +165,69 @@ declare namespace SauceNao {
 }
 
 declare namespace TraceMoe {
-  declare interface doc {
-    from: number;
-    to: number;
-    anilist_id: number;
-    at: number;
-    season: string;
-    anime: string;
-    filename: string;
-    episode: number;
-    tokenthumb: string;
-    similarity: number;
-    title: string;
-    title_native: string;
-    title_chinese: string;
-    title_english: string;
-    title_romaji: string;
-    mal_id: number;
+  // declare interface doc {
+  //   from: number;
+  //   to: number;
+  //   anilist_id: number;
+  //   at: number;
+  //   season: string;
+  //   anime: string;
+  //   filename: string;
+  //   episode: number;
+  //   tokenthumb: string;
+  //   similarity: number;
+  //   title: string;
+  //   title_native: string;
+  //   title_chinese: string;
+  //   title_english: string;
+  //   title_romaji: string;
+  //   mal_id: number;
+  //   synonyms: string[];
+  //   synonyms_chinese: string[];
+  //   is_adult: boolean;
+  // }
+
+  // declare interface response {
+  //   RawDocsCount: number;
+  //   RawDocsSearchTime: number;
+  //   ReRankSearchTime: number;
+  //   CacheHit: boolean;
+  //   trial: number;
+  //   limit: number;
+  //   limit_ttl: number;
+  //   quota: number;
+  //   quota_ttl: number;
+  //   docs: doc[];
+  // }
+
+  declare interface anilistData {
+    id: number;
+    idMal: number;
+    isAdult: boolean;
     synonyms: string[];
     synonyms_chinese: string[];
-    is_adult: boolean;
+    title: {
+      native: ?string;
+      romaji: ?string;
+      english: ?string;
+      chinese: ?string;
+    };
+  }
+
+  declare interface resultData {
+    anilist: anilistData;
+    filename: string;
+    episode: ?number;
+    from: number;
+    to: number;
+    similarity: number;
+    video: string;
+    image: string;
   }
 
   declare interface response {
-    RawDocsCount: number;
-    RawDocsSearchTime: number;
-    ReRankSearchTime: number;
-    CacheHit: boolean;
-    trial: number;
-    limit: number;
-    limit_ttl: number;
-    quota: number;
-    quota_ttl: number;
-    docs: doc[];
+    frameCount: number;
+    error: string;
+    result: resultData[];
   }
 }